+++ src/main/java/org/repozoo/commons/range/factories/LocalDateRange.java
@@ -1,4 +1,4 @@
-package org.repozoo.commons.range.factory.impl;
+package org.repozoo.commons.range.factories;
 
 import org.repozoo.commons.range.Range;
 import org.repozoo.commons.range.RangeFactory;
