+++ src/test/java/org/repozoo/commons/range/RangeSetTest.java
@@ -1,6 +1,6 @@
 package org.repozoo.commons.range;
 
-import org.repozoo.commons.range.factory.impl.IntRange;
+import org.repozoo.commons.range.factories.IntRange;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
